test(media): extract synthesis helpers to tests/common/mod.rs
WHY: pure-move refactor in preparation for slice 2 of T1 test
architecture. Existing tests/integration.rs is 615 LOC; new
edge-case tests pushed it over the 700-LOC Batch F/G soft-limit.
Per the convention, shared integration-test helpers live in
tests/common/mod.rs.

Zero behavioural delta — the 6 existing tests pass byte-for-byte.
Only change to helper bodies is adding `pub` visibility so test
binaries can `use common::{...}`.

The two #![allow] header attrs (`unreachable_pub`, `dead_code`) on
common/mod.rs are load-bearing per the Batch F/G workaround for
rust-lang/rust#46379 — every test binary compiles common/mod.rs
but only uses a subset of helpers.
